requirements.txt
Django==1.8
psycopg2
ipdb
pytest-django
PyQuery
django-tastypie
